refactor: movendo arquivo NotificationReceiver para pasta receiver
