--- app/src/main/java/com/example/safebyte/receiver/AlarmReceiver.kt
@@ -1,2 +1,0 @@
-package com.example.safebyte.receiver
-
